Move RestSharp tests to their own namespace
