+++ tests/Agent/IntegrationTests/IntegrationTests/RestSharp/RestSharpClientInstrumentationAsyncAwait.cs
@@ -9,7 +9,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.RestSharp
 {
     [NetFrameworkTest]
     public class RestSharpInstrumentationAsyncAwait : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
